Tricks4live: add/improve pages

+++ lib/entries/label.dart
@@ -1,5 +1,5 @@
 class Label {
-  BigInt id;
+  int id;
   String nameCN;
   String nameEN;
